requirements.txt
streamlit
tensorflow
requests
pydub
Pillow
numpy
librosa
soundfile
pyttsx3
